WIP: Trying to get records into Message
Currently running afoul of the borrow checker, which seems to think that
the borrowed slice that is taken by the record structs still belongs to
the local function when I try to return the completed Message struct.
Not really sure what to do about this yet. Maybe need to build the
record structs from methods on Message?