WIP: Major overhaul to using the bytes crate
Using the bytes crate brings a lot of useful tools to greatly simplify
processing the data coming off the network, as well as building the data
I'm sending on the network.

Unfortunately at the moment handling compressed labels is broken, as the
bytes crate has made significant changes to how I work with the data and
I haven't quite figured out how to resolve this yet.