Using Bytes/BytesMut structs from bytes crate
This will need benchmarking to see if they are in fact more performant
in any way. The API doesn't seem to be giving us any solid advantages
over what we were doing before, other than the flexibility that we
could, if we choose, define buffer size at runtime instead of compile
time.